Use custom cmake variable for project name and target (#3)
CMAKE_PROJECT_NAME is supposed to be the topmost project which means
that it will propagate down to dependencies if those are included via
FetchContent. This means that Hexi cannot be included in a project via
FetchContent because the CMAKE_PROJECT_NAME is taken from the project it
is being included by.

Co-authored-by: Robin Küster <robin@kuester.me>