Add adaptors for explicit control over string representation (#14)
Provides adaptors for explicitly specifying how strings should be handled with being serialised and deserialised.

raw: write the data exactly as in
null_terminated: ensure the string is written with a null terminator - strings with embedded null bytes are not allowed (but is only enforced in debug) as retrieving such a string would stop at the first embedded null. This is user error is mostly semantics.
prefixed: prefixes the length before the string, no terminator is written.
prefixed_varint: prefixes the length, encoded as a varint, no terminator is written. This is more compact but slower than prefixed

All adaptors can be used both as input and output, with the exception of raw(). Retrieving strings written using raw requires the user to know specifics about the string content (i.e. whether it has embedded nulls or its length), so .get() should be used.